:fist_raised: Changed navigation structure
