+++ pride/src/components/FloatingTabBar/FloatingTabBar.tsx
@@ -62,7 +62,7 @@ const styles = StyleSheet.create({
     right: 0,
     margin: ms(margin),
     height: ms(height),
-    borderRadius: 100,
+    borderRadius: ms(height * 2),
     backgroundColor: backgroundColor,
     flexDirection: 'row',
     overflow: 'hidden'
