--- pride/src/navigation/ProfileNavigator/index.tsx
@@ -1,3 +1,0 @@
-import ProfileNavigator from './ProfileNavigator';
-
-export default ProfileNavigator;
