+++ pride/src/navigation/TabNavigator/index.tsx
@@ -1,0 +1,3 @@
+import TabNavigator from './TabNavigator';
+
+export default TabNavigator;
